Question Answers
Animation Practice Answers:
(1) Code line 11 draws the blades of grass ( line(x, height-10, x + random(-10, 10), height-10-random(h))

(2) Line 23 makes the "lawnmower" come by. It comes by when the grass reaches a certain height, the first time when it fills the canvas and the rest of the time it comes by when its above the middle of the canvas.

(3) the H variable represents the height of the blades of grass 

(4) they set the range of value of the hue, saturation and brightness of the colors

(5) The first -10 controls the way the grass is on the ground, changing it to -20 makes it higher off the ground. The height-10 random(h) controls the height of the grass and lets the lawnmower come later after the grass reaches a certain height